[FEAT] Split the colour model out of AppearancePreferences
AppearancePreferences.kt held eleven top-level types across 518 lines, well
past the few hundred the Kotlin conventions suggest, and the colour model
was the worst fit: CutoutColor is used by 18 files and DynamicRole by 12,
yet both were imported from a file named after appearance storage.

- data/CutoutColor.kt: CutoutColor, DynamicRole, GradientDirection,
  ColorSpec, CutoutFill
- data/ActionButtonStyles.kt: ActionButtonStyle, ActionButtonAlignment,
  SentAlignment, ReplyInputStyle
- AppearancePreferences.kt keeps AppearanceSettings and its store, 285 lines

Pure code motion. Everything stays in the data package, so no import
changed anywhere, and no serialization format or DataStore key was touched.
